examples/onelake-local/requirements.txt
azure-ai-evaluation==1.3.0
azure-identity==1.20.0
azure-storage-file-datalake==12.18.1